Remove KeyValueEntityTracker
Force apps to write their own logic so that logic is explicitly part of the app and cannot be misconstrued to do something else.

Issue #2176
